+++ doc/source/code/tut5.py
@@ -12,10 +12,10 @@ netParams['probLengthConst'] = 200.0 # propagation velocity (um/ms)
 
 ## Population parameters
 netParams['popParams'] = []  # list of populations - each item will contain dict with pop params
-netParams['popParams'].append({'popLabel': 'ES', 'cellType': 'PYR', 'numCells': 10, 'xnormRange': [0,1.0], 'ynormRange': [0,0.5], 'cellModel': 'HH'}) 
-netParams['popParams'].append({'popLabel': 'IS', 'cellType': 'BAS', 'numCells': 10, 'xnormRange': [0,1.0], 'ynormRange': [0,0.5], 'cellModel': 'HH'}) 
-netParams['popParams'].append({'popLabel': 'EM', 'cellType': 'PYR', 'numCells': 10, 'xnormRange': [0,1.0], 'ynormRange': [0.5,1.0], 'cellModel': 'HH'}) 
-netParams['popParams'].append({'popLabel': 'IM', 'cellType': 'BAS', 'numCells': 10, 'xnormRange': [0,1.0], 'ynormRange': [0.5,1.0], 'cellModel': 'HH'}) 
+netParams['popParams'].append({'popLabel': 'ES', 'cellType': 'PYR', 'numCells': 100, 'xnormRange': [0,1.0], 'ynormRange': [0,0.5], 'cellModel': 'HH'}) 
+netParams['popParams'].append({'popLabel': 'IS', 'cellType': 'BAS', 'numCells': 100, 'xnormRange': [0,1.0], 'ynormRange': [0,0.5], 'cellModel': 'HH'}) 
+netParams['popParams'].append({'popLabel': 'EM', 'cellType': 'PYR', 'numCells': 100, 'xnormRange': [0,1.0], 'ynormRange': [0.5,1.0], 'cellModel': 'HH'}) 
+netParams['popParams'].append({'popLabel': 'IM', 'cellType': 'BAS', 'numCells': 100, 'xnormRange': [0,1.0], 'ynormRange': [0.5,1.0], 'cellModel': 'HH'}) 
 netParams['popParams'].append({'popLabel': 'background', 'rate': 10, 'noise': 0.5, 'cellModel': 'NetStim'})
 
 ## Cell property rules
@@ -44,16 +44,16 @@ netParams['synMechParams'].append({'label': 'inh', 'mod': 'Exp2Syn', 'tau1': 0.1
 ## Cell connectivity rules
 netParams['connParams'] = [] 
 
-netParams['connParams'].append({'preTags': {'cellType': 'PYR'}, 'postTags': {'popLabel': ['PYR','BAS']},  #  E -> all
-	'probability': '0.5*post_normy', 		# probability of connection
-	'weight': 0.00, 						# synaptic weight 
-	'delay': 'dist3D/propVelocity',			# transmission delay (ms) 
+netParams['connParams'].append({'preTags': {'cellType': 'PYR'}, 'postTags': {'cellType': ['PYR','BAS']},  #  E -> all
+	'probability': '0.1*post_ynorm', 		# probability of connection
+	'weight': 0.01, 						# synaptic weight 
+	'delay': 'dist_3D/propVelocity',			# transmission delay (ms) 
 	'synMech': 'exc'})   					# synaptic mechanism 
 
-netParams['connParams'].append({'preTags': {'cellType': 'BAS'}, 'postTags': {'popLabel': ['PYR','BAS']},  #  I -> all
-	'probability': '1*exp(-dist3D/probLengthConst)', 	# probability of connection
+netParams['connParams'].append({'preTags': {'cellType': 'BAS'}, 'postTags': {'cellType': ['PYR','BAS']},  #  I -> all
+	'probability': '1*exp(-dist_3D/probLengthConst)', 	# probability of connection
 	'weight': 0.00, 									# synaptic weight 
-	'delay': 'dist3D/propVelocity',						# transmission delay (ms) 
+	'delay': 'dist_3D/propVelocity',						# transmission delay (ms) 
 	'synMech': 'inh'})   								# synaptic mechanism 
 
 netParams['connParams'].append({'preTags': {'popLabel': 'background'}, 'postTags': {'cellType': ['PYR', 'BAS']}, # background -> all
@@ -66,7 +66,7 @@ netParams['connParams'].append({'preTags': {'popLabel': 'background'}, 'postTags
 simConfig = {}
 simConfig['duration'] = 1*1e3 			# Duration of the simulation, in ms
 simConfig['dt'] = 0.025 				# Internal integration timestep to use
-simConfig['verbose'] = False  			# Show detailed messages 
+simConfig['verbose'] = True  			# Show detailed messages 
 simConfig['recordTraces'] = {'V_soma':{'sec':'soma','pos':0.5,'var':'v'}}  # Dict with traces to record
 simConfig['recordStep'] = 1 			# Step size in ms to save data (eg. V traces, LFP, etc)
 simConfig['filename'] = 'model_output'  # Set file output name